fix(contracts+cli): address 4th Gemini review — OOB, truncation, cell capacity, efficiency
proposal-anchor: validate_pblk minimum length raised from 38 to 39 — the
v0x01 path accesses data[38] as id_len, so len==38 caused an out-of-bounds
panic. The v0x02 branch enforces exact length separately and is unaffected.

governance-lock: merkle_leaf_index comparison no longer truncates u32 to u16.
Changed 'entry.merkle_leaf_index as u16 >= header.validator_count' to
'entry.merkle_leaf_index >= header.validator_count as u32' — a malicious
validator could submit index 65536 (truncates to 0) to bypass the check.

execute.ts + gui-server.ts: proposal change and extraTreasuryOutputCapacity
merged into a single treasury output. A separate extraTreasuryOutputCapacity
cell could fall below the 125 CKB minimum for treasury-lock's 64-byte args,
causing the node to reject the transaction.

treasury-lock: precompute governance-lock hash once before the output loop
instead of loading the full 54-byte lock script of every output cell.
Removed is_governance_lock() — now replaced by direct hash comparison.

+++ contracts/governance-lock/src/main.rs
@@ -568,7 +568,7 @@ fn program_entry() -> Result<(), i8> {
         if entry.vote != 1 {
             return Err(ERR_INVALID_WITNESS);
         }
-        if entry.merkle_leaf_index as u16 >= header.validator_count {
+        if entry.merkle_leaf_index >= header.validator_count as u32 {
             return Err(ERR_SIG_VERIFICATION);
         }
         if seen.iter().any(|pk| pk == &entry.pubkey) {
